Added custom version of the function ChangeTableSQL for PostgreSQL in the file "datadict-postgres.inc.php". This should add transparent database update support when using ADODB AXMLS. Please note the following:
The function is not to be confused with a general implementation. Do not be tempted to replace it with the same function in "adodb-datadict.inc.php".
The transparent database update for PostgreeSQL has the same limitation that would exist for other databases using ADODB AXMLS.
The function only supports PostgreSQL version 8 and up at the moment.
No attention has been paid to performance yet when implementing the function.

Referencing request "https://github.com/ADOdb/ADOdb/issues/124".