Towrads consistancy of the table name normalization behavior.
The function "ADOConnection::MetaColumns" has a normalization parameter that allows one to disable case normalization, which works by either converting to uppercase or lowercase. However other functions of the same logical group such as "ADOConnection::MetaColumnNames" have no such parameter and when they make the call to "ADOConnection::MetaColumns" they set the parameter themselves. This makes external code that relies on MetaColumns as well as MetaColumnNames for example buggy, unless the external code does not set the normalize parameter on MetaColumns. This is not a bug in the ADODB code, but a bug in the ADODB algorithm.

This is the first of many steps to come towards fixing this issue. This commit should be fully backward compatible. The syntax of backquotes introduced by ADODB Data Dictionary will be integrated along with a new quotation, "<" and ">" to indicate case sensitivity. The following is from the comments of the new function ParseIdentifierName():

If identifier name is identifierName, the following can be passed:
    FORM 1: identifierName = To keep default old behavior.
    FORM 2: `identifierName` = To quote identifierName with identifier or string quotation. Normalization is still done based on default behavior.
    FORM 3: <identifierName> = To keep identifier name without normalization. Some drivers will normalize regardless.
    FORM 4: `<identifierName>` = To quote identifierName with identifier or string quotation and keep identifier name without normalization. Some drivers will normalize regardless.

Please refer to https://github.com/ADOdb/ADOdb/issues/129 for more details, and the comments of the new function ParseIdentifierName().