Towards consistancy of the table name normalization behavior. (issues : 129)
Ammended MetaColumns, MetaColumnNames and MetaIndexes. Commit to proposal branch might be done later.

Some minor fixes to the MetaColumns on some drivers. (issue : 133) No proposal for this part of the commit is currently intended.

Achieved with this commit:
- Full adherence, when required, to case sensitivity of Table, Columns and index names on most, if not all, databases when using ADODB AXMLS.
- Better stability between the three libraries, ADODB, ADODB Data Dictionary, and ADODB AXMLS.
- ADODB AXMLS support for PostgresSQL.
- Layed out the road for future work in integrating ADODB's normalization feature into ADODB Data Dictionary and ADODB AXMLS.

Testing was done on the drivers MYSQL, MYSQLI, PDO-MYSQL, ODBC MSSQL, and POSTGRES9.