Towards full integration with ADODB Data Dictionary.
The PDO driver code follows an anti pattern breaking proper class inheritance. The class ADODB_pdo creates instances within its code of classes inheriting from it representing specific PDO drivers, such as the ADODB_pdo_oci class, saves them in a local class variable, and manually tries to mimic the inheritance structure. This anti pattern was already proven a hindrance during the integration with ADODB Data Dictionary work. The code has been fixed to follow proper inheritance.

As a side a effect of the above, one can longer use the PDO driver by using the string "pdo" when calling ADONewConnection(), but instead must use specific PDO based driver names such as "pdo_pgsql".

Another side effect is that PDO no longer supplies a database driver, whose class was "ADODB_pdo_base" for when no connection was established, but no exception was thrown either. We are unsure if this was a possible situation. The driver was also used as a default driver when a connection was established, but no specialized PDO based driver was found. The code has been kept, but is currently unusable, and might be re introduced later if needed.

Dormant code was found in the drivers "adodb-sqlite.inc.php" and "adodb-sqlite3.inc.php" which is code relevant to PDO and appears to have been the side effect of a copy and paste operation. The code has been removed. Note that the code would no longer be found in the PDO drivers because it was part of the anti pattern.