Towards full integration with ADODB Data Dictionary. Step 3: ADOConnection::$concat_operator, ADOConnection::sysDate, ADOConnection::sysTimeStamp, ADOConnection::$sysUTimeStamp, ADOConnection::SQLDate().
Moved the afore mentioned functions and variables into ADODB Data Dictionary. Code should be fully backward compatible.

The variable ADOConnection::$sysUTimeStamp was deleted because no other use of it was found anywhere in the entire library.

Some traces of the above remain in the sqlite drivers due to their use of PHP functions inside sql statements, "adodb_date2()" and "adodb_date()". These traces shall be moved over once the said behavior is implemented in the sqlite driver for PDO.

Functions ending with 'SQL' inside the ADODB Data Dictionary classes, are functions meant to produce SQL and should generally be driver independent. This means they should work without a valid driver. Unfortuanely, some instances of the new function introduced, FormatDateSQL(), make calls to the qstr() function which is inside the ADODB drivers themselves.

The work on this commit revealed how much the drivers came to diverge over the years. We hope with this work that the community finds it easier to make changes and fixes that benefit all drivers due to the factored code.

A minor fix to the driver "adodb-pdo_pgsql.inc.php" slipped through during this update, but should be no hazard. The commit of the fix was to be postponed till the integration is done.