Towards full integration with ADODB Data Dictionary
Fixed new functions that we introduced with names ending with SQL in ADODB Data Dictionary not returning an array.

Finished integration of ADOConnection::RowLock(). This work has shown us the extent of the discrepancy that exists between implementations in different drivers. For example, some drivers would accept a single table name, while others multiple. The main ADODB code and documentation suggests single table name. Due to this, and a number of other issues, this is the first commit in the integration work that is likely to have introduced behavioral changes. Much analysis has been done to ensure this is not the case, but we are still uncertain.

Traces of ADOConnection::RowLock() still exist in some drivers due to the discrepancies talked about above. In some instances the traces remain due to existence of comments that describe the changed behavior.

The first parameter of RowLock() expects a table name per documentation. This makes it a candidate for the quotation syntax introduced by ADODB Data Dictionary, and built upon by us. However, because all drivers, except one, agree it is tables instead of a single table name, the current definition that we worked from has been Table Name SQL Segment. Perhaps in the future we could change the definition to comma delimited list of table names and integrate the afore mentioned syntax. However careful examination of all SQL variations has to be made first.

A minor fix to BeginTrans() for the adodb-mysql.inc.php driver has been added to ensure compliance to apparent specification.

RowLock() in the driver adodb-sybase.inc.php probably never worked correctly due to reliance on the class variable '_hastrans' which is not defined or exists anywhere in the entire library. Fixes were made assuming the hasTransactions class variable.