Towards full integration with ADODB Data Dictionary step 3: CreateSequence() (stage 1)
Factoring the ADOConnection::CreateSequence() function has proven to be exceedingly difficult. This is the reason why this factoring is being done in stages.

The driver adodb-ads.inc.php was found to ignore the start id parameter. This was not fixed.

Fixes in core ADOConnection::CreateSequence() to ensure solid boolean return. This should not affect backward compatibility.

Fixes in the adodb-odbc.inc.php driver to ensure the function signature of "DropSequence()" is compliant

Full removal of the create sequence code in the function GenID in the drivers adodb-mysql.inc.php and adodb-mysqli.inc.php was not achieved yet due to the unusual variation of error checking done by those drivers. Unusual compared to other drivers.

Two behavioral changess were introduced into the "adodb-ado_mssql.inc.php" and "adodb-mssql.inc.php" drivers. (1) The function CreateSequence() continues to create a transaction, but no longer creates an explicitly named one. (2) The function CreateSequence() continues to create a transaction but only under the same conditions that bound the function BeginTrans(), namely the class variable "transOff", while before this commit the function would always start a transaction. However the function continues to do what it is supposed to do even without the transaction.

A potential bug was fixed in the GenID2008() function in the driver "adodb-mssqlnative.inc.php" that would have caused it to return the wrong id, if the sequence was created during the same call. The driver was also found to be significantly divergent from other mssql drivers, and non conformant to apparent specification in a number of places. Careful care should be used when factoring this driver.

A behavioral change was introduced into the drivers "adodb-ado_mssql.inc.php", "adodb-mssql.inc.php", and "adodb-mssqlnative.inc.php" that can cause the function GenID to create two transaction instead of one. However we think the benefit of the code factoring that introduced this far outweighs the cost.

Two parasitic class variables were found, "$_genSeqSQL" and "$_genSeq2SQL", in the drivers that should be eventually removed.

There still remain instances where the function CreateSequence() does not return a solid boolean.