Towards full integration with ADODB Data Dictionary step 3: CreateSequence() (stage 2)
A fix to our previous commit for the SQLite drivers.

Fixes to the function signatures of CreateSequence() and DropSequence() in a number of drivers to comply with specification.

The parasitic class variable $_genSeqSQL2 is now fully removed.

A minor change to the behavior of CreateSequence() in the "adodb-ads.inc.php" and "adodb-mssqlnative.inc.php" drivers in the way they used to print debug and error messages. This should not affect backward compatibility.

The function CreateSequence() in the drivers "adodb-ado_mssql.inc.php", "adodb-mssql.inc.php" and "adodb-mssqlnative.inc.php" (CreateSequence2008) appears to could have been used to reset the sequence to a particular number. This is against apparent specification and documentation. This usage is no longer possible.

Traces of the CreateSequence() function still exists in the “adodb-mssqlnative.inc.php”, “adodb-odbc.inc.php” and “adodb-odbtp.inc.php” drivers due to unusual implementations. Factoring for the “adodb-mssqlnative.inc.php”, “adodb-odbc.inc.php” should still be possible and is intended to be done in the future, but we are unsure about “adodb-odbtp.inc.php”.

The parasitic class variable $_genSeqSQL has been removed as a parasite, but continues to be defined and used in the drivers “adodb-odbc.inc.php” and “adodb-odbtp.inc.php”. It is likely to be removed in the future.