Towards full integration with ADODB Data Dictionary step 3: DropSequence(), GenID() (stage 1), CreateSequence() (stage 3)
Traces of DropSequence() continue to exist in the "adodb-odbc.inc.php" and "adodb-odbtp.inc.php" drivers. Factoring out could not be done at this point, but should likely be doable after the work on GenID() is finished.

Code analysis of GenID() revealed that the function must fail safely on the server level, and on the database level. This means the function should not fail a running transaction. The function explicitly disables the functionality of the StartTrans() function family, which should not be confused by the BeginTrans() function family. This and further analysis revealed that the function CreateSequence() must not be transactional. Our previous commits turned the function transactional through the factored algorithm we introduced. The exceptions that lead to this were mssql drivers. It appears that the mssql drivers had a transactional implementation that was meant to protect from a SQL Server specific behavior related to automatic rollback at the database level. The behavior was implemented outside of ADODB's transactional mechanisms such as to avoid failure with the two function families mentioned earlier. We are unsure if this the case, but for the time being the behavior was restored.

Due to the above, all behavioral changes introduced to the "adodb-ado_mssql.inc.php", "adodb-mssql.inc.php" and "adodb-mssqlnative.inc.php" drives in the previous two commits have been undone, apart from the following statement in the previous commit's notes which continues to be true :

	The function CreateSequence() in the drivers "adodb-ado_mssql.inc.php", "adodb-mssql.inc.php" and "adodb-mssqlnative.inc.php" (CreateSequence2008) appears to could have been used to reset the sequence to a particular number. This is against apparent specification and documentation. This usage is no longer possible.