Towards full integration with ADODB Data Dictionary step 3: GenID() (stage 3)
The driver "adodb-ads.inc.php" no longer outputs an error message on error during a run of ADOConnection::GenID(). This is not part of apparent specification.

Due to the implementation of "ADOConnection::_insertid()" in the driver "adodb-mysql.inc.php", ADOConnection::GenID() no longer uses mysql_insert_id(), and instead uses a sql statement, to retrieve the value. However, the new code is written such as if ADOConnection::_insertid() is returned to use mysql_insert_id(), the behavior is restored.

Traces of ADOConnection::GenID() continue to exist in some drivers due to the existence of a loop that continuously tries to do the GenID() operation for a number of times on failure before finally failing completely. These particular traces however continue to rely on the core ADOConnection::GenID(), and hence continue to follow specification.

Work is still required on the ODBTP drivers, and SQL Server drivers.

Warning: The code should be considered unstable for the SQL Server database after this commit, regardless of driver. This should be resolved on the next commit.