Towards full integration with ADODB Data Dictionary step 3: DropSequence(), GenID() (stage 4), CreateSequence() (stage 4)
MSSQL drivers that were not returning a positive number, or zero on error after a call to ADOConnection::GenID() are now doing so. Returning otherwise was against apparent specification.

A number of bugs were fixed by the factoring process in the driver "adodb-mssqlnative.inc.php".

The public variable "$sequences" was removed from the  ADOConnection implementation in the driver "adodb-mssqlnative.inc.php" to allow factoring. The benefit of the factoring should exceed the cost of the performance loss.

In some implementations errors were outputted on error during calls to one of the three functions addressed by this commit. These are against apparent specification and have been removed whenever they hindered the factoring process.

The MSSQL drivers can now be considered stable in matters pertaining to sequence generation, annulling the instability warned about in a commit two weeks ago.