Code reduction: ADOConnection::fetchMode.
There existed large dependency on  developers to treat ADOConnection::fetchMode and ::ADODB_FETCH_MODE correctly resulting in inconsistency and growing divergence in what seemed to be newer code. The non intuitive behavior of ADOConnection::SetFetchMode() seems to have exacerbated the matter. This is the first in a number of commits intended to be to fix the issue.

Note that this reduction is not a sufficiently equivalent transformation. A number of places in the pre reduced code showed apparent contradiction between intent and algorithm. Forced reduction by its nature confirms code to the more apparent. This will likely fix a number of hidden bugs, and break a number of ad hoc fixes for said bugs.