Code Reduction: ::ADODB_FETCH_MODE.
Warning: This commit changed many implementations of ADOConnection::MetaColumns to obey the fetch mode set in the current driver instance, where before they would obey the ::ADODB_FETCH_MODE variable before, which is an error per apparant specification. ::ADODB_FETCH_MODE is only the default value of the fetch mode that an instance of ADOConnection takes if no fetch mode is set for the instance explicitly.