Removing the distrupting function ADORecordSet::_updatefields().
It appears that there is a misunderstanding on what the variable ADORecordSet::$fields, which is a variable that must hold the raw data as it comes from the driver, means. Incidentally, whether the variable must follow a minimal structure as well is not clear at this point. In any case, we think that there has been a desire to make certain functions work on the expected character case, functions that rely on  ADORecordSet::$fields to work, and that desire led to the current incomplete solution, or solutions, involving ADORecordSet::_updatefields(). One such function is likely ADORecordSet::Fields().

This is one step in a number of steps to solve the pertaining issue. As of this commit, this fork is not stable.