Amending ::ADODB_FETCH_MODE
Warning: ADODB_FETCH_DEFAULT now resolves to ADODB_FETCH_BOTH. On most drivers this should not be an issue unless you are not using the constants in your code. On at least one driver, adodb-sybase.inc.php, this changes from ADODB_FETCH_ASSOC to ADODB_FETCH_BOTH but should not be a problem in general. If you are using a driver that only offers ADODB_FETCH_NUM or only offers ADODB_FETCH_ASSOC you have to change your code from using ADODB_FETCH_DEFAULT to one of the other constants.