Towards amending the constant ADODB_ASSOC_CASE.
Removed the constant ADODB_ASSOC_CASE.

Amended definition of ADODB_ASSOC_CASE_xxx constants.

Fixed apparent incorrect association of the ADODB_ASSOC_CASE_xxx constants with the ADODB php extension.

Repaired ADORecordSet::GetRowAssoc() default value. Users who want a different default should explicitly pass a default.

It was deliberated that the constant ADODB_ASSOC_CASE definition be changed to become "default value for functions that support ADODB_ASSOC_CASE_xxx constants", but because this definition can introduce bad coding habits, and because the constant is new, and the constant is not a valid solution to the problem it was introduced for, which is a problem already fixed in this fork, and only a single function, or two roughly in the main branch, uses the constants, this idea was eventually dropped.