Postliminary Commit.
The function ADORecordSet::FetchField() uses a -1 as a parameter default to indicate to algorithm that the variable was not passed. When the variable is not passed, the behavior is driver dependent. Some drivers will fail, while some will pass the next field, and one or few others will pass the first field. This is not a bug. Using this method without passing a variable is highly discouraged however.

ADORecordSet::FetchField() for PDO drivers have been changed to return the first field when no variable is passed. This is an extension of the previous community commit's apparent desire for the pdo_sqlsrv driver.

Repairing ADORecordSet::FetchField() for PDO drivers. PDOStatement::getColumnMeta() is zero based. Thanks for the community previous commit for pointing this out for the pdo_sqlsrv driver.