Official commit "876d51b73a5efa089e4cb12589de242e4f24b4c2", dated Jul 19, 2018 made fixes to ADOdb Active Record to honor the global variable "$ADODB_QUOTE_FIELDNAMES". The official documentation, at
	https://adodb.org/dokuwiki/doku.php?id=v5:reference:adodb_quote_fieldnames
has a definition for "$ADODB_QUOTE_FIELDNAMES" that seems to suggest that it affects all parts of ADOdb. However the page at
	https://adodb.org/dokuwiki/doku.php?id=v5:reference:adodb_quote_fieldnames
suggest that it  is only meant to affect the "autoExecute()" function, which is ADOConnection::AutoExecute(), which in turn affects ADOConnection::GetInsertSQL() which affects _adodb_getinsertsql(), and affects ADOConnection::GetUpdateSQL() which affects _adodb_getupdatesql(). This definition is confirmed by an old copy of the ADODB manual made in the year 2014 back when the manual was a single page. The intended effect on the other two functions of ADOConnection is confirmed by old comments in the code. This contradicts the entire premise of the reason for the commit, and is already in line with what the ADOdb code was doing.

When merging of that commit was addressed by this fork, this contradiction was not known. Instead, in this fork, the class static variable, ADODB_Active_Record::$_quoteNames was noticed, whose meaning was also changed by the above official commit. In this fork, the variable's previous meaning was taken to have the old intended behavior pertaining to this matter for the ADODB_Active_Record class. This understanding itself is not incorrect. However, because the above official commit sought to permeate the effect of "$ADODB_QUOTE_FIELDNAMES" per the wrong understanding throughout ADODB_Active_Record, this fork sought to retain that intentions as much as possible, but instead use the correct variable ADODB_Active_Record::$_quoteNames and its original meaning to retain backward compatibility and original intent, through the function ADODB_Active_Record::_QName().

Now that the above contradiction is known, the changes in this fork should be reverted. Further more it now appears that ADODB_Active_Record::_QName() was part of an original unfinished work introduced between versions 5.08a and 5.09.

The active record mechanism is expected to hide the SQL details from you, and hence, it can be reasoned that the work of using ADODB_Active_Record::_QName() should be kept, even though the original reasons behind doing it were incorrect. However, the official commit that this postliminary commit is adressing, which is different form the official commit mentioned above, proved another problem. The ADODB_Active_Record relies for one of its functions, ADODB_Active_Record::Replace(), on the ADOConnection::Replace() function which does not permeate the behavior of ADODB_Active_Record::_QName(), or in other words, the whole ADODB_Active_Record::$_quoteNames related semantics, and it can not be expected to do so, otherwise it would be ill defined. The official commit that this postliminary commit is adressing is a path in that wrong direction, and can be expected to break backward compatibility.

Given that an active record mechanism is expected to hide the SQL details, it has been decided to indeed keep the work on the ADODB_Active_Record::_QName(). As for the ADODB_Active_Record::Replace() function, a new mechanism function had to be introduced to break its coupling from ADOConnection, and plug any leaking abstraction. However, the variable ADODB_Active_Record::$_quoteNames will still be kept for backward compatibility and other reasons (see below), but note that given ADOdb's code convention, that variable appears private.

For future work, note the following:
- There are two issues pertaining to identifiers in SQL. The first is case sensitivity, and the second is clarity such as an identifier with value equal to a reserved keyword. How these problems can be addressed in each database engine depend on that engine, and client code convention. There are two main mechanisms, character casing and quoting, and they do not mean the same thing in every engine. For example, in some engines, quoting means that the identifier is case sensitive, and so is not simply for clarity. The ADOdb code must not address the semantics of casing and quoting in general, but in the case of the active record implementation, this is necessary. Currently quoting is used in the active record mechanism with the intent to achieve both clarity and case sensitivity, and it is hoped that this effectively holds in all database engines, but it is not proven and the keeping of ADODB_Active_Record::$_quoteNames. As for other parts of the library, ADOdb may only promise casing or quoting, but not their semantics.