Postliminary Commit.
- In this fork, function names must retain their old capitalization.

- Fixed the corruption of ADOConnection::fetchMode by the new DB2 driver when modifying the fetch mode for internal purposes. Corruptions occured for a number of different reasons, one being the ignoring of one of the states of ADOConnection::fetchMode, a state indicating that the value should be retrieved from the current value of $ADODB_FETCH_MODE.

- Fixed a bug ADOConnection::ServerInfo() for the legacy DB2 driver, and the new driver of DB2, related to the fetch mode.

- The new driver of DB2 contains specification comments for virtual methods first defined in the base ADOConnection driver. These comments must be removed, but were retained for the time being to avoid much differences from the main fork. The comments, however, remain non authoritative on the specification of said methods in this fork.

- Configuration related to casing, which is related to normalization, does not pertain, and must not affect, returns from databases. These only affect inputs to databases. Whether the constant ADODB_ASSOC_CASE, which this fork removed, or the global variable ::$ADODB_ASSOC_CASE. Violations were removed from the new db2 driver. For similar reasons, ADODB_db2::getMetaCasedValue() was removed from the new DB2 driver.

- In this fork, instances of ADORecordSet, must be passed the fetch mode from the ADOConnection drivers creating them.

- In this fork, ::$ADODB_FETCH_MODE must not appear any where in the driver code.

- ADODB_db2::setTableCasing(), ADODB_db2::getTableCasing(), ADODB_db2::getTableCasedValue() and the constants ADODB_db2::TABLECASE_LOWER, ADODB_db2::TABLECASE_UPPER, and ADODB_db2::TABLECASE_DEFAULT, were removed from the new DB2 driver. Refer to the specification of ADODB_DataDict::ParseIdentifierName() defined in this fork.

- The excessive private function, ADORecordSet_db2::processMoveRecord(), was removed.

- The method ADODB_DataDict::ChangeTableSQL(), in the DB2 data dictionary, did not correctly set the fetch mode on the underlying ADOConnection driver before calling ADODB_DataDict::MetaColumns().

- Fixed a bug in ::ADODB_Error_Handler() for the case where the $sql parameter is an array.

- Fixed a minor bug related to the work on ADORecordSet_array done by this fork.

- WARNING: CommitTrans
	https://github.com/ADOdb/ADOdb/commit/fe52d38092f768e95fb67438527beda276ec6ae3
appears to have sought to improve on a previous commit						https://github.com/ADOdb/ADOdb/commit/60c48af0bf265a200070466aa4de97dd5126ed06
which sought to remove the usage of PHP's $php_errormsg variable. This second commit, except for possible minor errors in the logic of setting ADOConnection::_errorMsg, was valid, except for the issue that PHP's ::error_clear_last() was not available untill PHP v7. The first commit sought to improve on the second commit's treatment of the array returned by PHP's ::error_get_last(), which the second commit introduced usage of instead of PHP's $php_errormsg, by introducing two functions, ADOConnection::resetLastError() and ADOConnection::getChangedErrorMsg(), which were not equivilant to the code they were replacing, making them ill defined. The first commit, however, was based on an explicit assumption that code that was setting $php_errormsg to empty string was doing so to detect changes to it. The commit that this postliminary commit is addressing proved that assumption is wrong. Similar code existed in the old driver that the above first commit simply removed because code checking for changes to $php_errormsg did not exist in said code after setting the variable to empty string.

The above two commits made it to the official commit						https://github.com/ADOdb/ADOdb/commit/bd898bf510431e2f8bce2469d188821e200af29f
It was unclear, at least to this author, whether $php_errormsg is officially writable, and whether setting it to an empty string is equivalent to calling PHP's ::error_clear_last(), and whether setting it to a non empty string has side effects, and whether those side effects have an explicit equivalent in newer PHP. Or addressed differently, why the old code would check if $php_errormsg is set before assigning an empty string to it, and why the first commit above did not simply clear the error using either ::error_clear_last() when it is available, or $php_errormsg otherwise. Because of this, this postliminary commit does not address the re introduction of PHP's $php_errormsg variable into the code, and in particular the new db2 driver. In the old db2 driver, db2legacy, which this fork introduced, likely safe code, per this author's knowledge, was introduced to handle clearing PHP's $php_errormsg when it appears that the intention of the code to has been as such, and not to has been to detect changes to $php_errormsg.