Postliminary Commit.
ADODB_pdo::$dsnType was made protected, and remaining usage of it in the ADODB code outside of the PDO drivers has been removed. In this fork, ADOConnection::$databaseType is normalized.

This fork continues to support PHP 5 (roughly 5.2) and up. Calls to ::assert() in "adodb-compress-bzip2.php" and "adodb-compress-gzip.php" have been removed.