Postliminary Commit.
Column order is crucial to the operation of ::_adodb_getmenu() ::_adodb_getmenu_gp(), and by extension to the behavior of ADORecordSet::GetMenu(), ADORecordSet::GetMenu2() and
ADORecordSet::GetMenu3(). The pertaining commit potentially damages the ordering.

Excess code introduced by the pertaining commit was kept to keep differences with the official fork minimal.

The existence of ADORecordSet::GetMenu(), ADORecordSet::GetMenu2() and
ADORecordSet::GetMenu3() is ill defined. The functions are kept for backward compatibility.

Note: In this fork, ADORecordSet::fetchMode is public but read only to external code.