Postliminary Commit.
Changing a variable to a constant changes its semantic, and therefore it is generally ill defined, and it is indeed in this case with the variable ::$ADODB_ASSOC_CASE.

Continue using ::$ADODB_ASSOC_CASE to affect the active record, if using this fork.

ADODB_ASSOC_CASE and ::$ADODB_ASSOC_CASE were considered parasitic, and the constant ADODB_ASSOC_CASE was already removed in the past.