Postliminary Commit
The existence of of html generation functions, such as ADORecordSet::GetMenu(), ADORecordSet::GetMenu2() and ADORecordSet::GetMenu3() is ill defined. The newly added function, ADORecordSet::getMenuGrouped(), was removed.

The pertaining commit is already based on an incorrect premise, but corrects an error introduced by an earlier official commit that damaged the column order, which is crucial to the operation of ::_adodb_getmenu() ::_adodb_getmenu_gp(), but the pertaining new code continues to be ill defined. However, factoring done to both of these functions was kept.