Postliminary Commit
The fix in the pertaining commit is not required because of the normalization and architecture correction work that was done in this fork. Further more, the newly added code in the pertaining commit was deduced to be effectively VERBATIM from the "adodb-mysql.inc.php" driver in the main fork. Hence, backward compatibility, or lack of, for the functions ADOConnection::GenID(), and ADOConnection::CreateSequence() remains the same as that of "adodb-mysql.inc.php" per the work that was already done in this fork.