Postliminary Commit
Much care must be exercised when adding comments to the base classes when non was for a long time, because such comments can be assumed authoritative. I fixed what I can from these newly added comments, but in this fork, the code itself remains the authoritative source and in order of its permeation and age.