Postliminary Commit
Much care must be exercised when adding comments to the base classes when non was for a long time, because such comments can be assumed authoritative. I fixed what I can from these newly added comments, but in this fork, the code itself remains the authoritative source and in order of its permeation and age.

Care must be taken with types "mixed" and "mixed[]". These should not be used in general, and instead an exhaustive list of types must be provided. For example, a parameter specified as "array" only means that it is an array. It does not specify the type of the contained elements, but a parameter specified as "mixed[]" explicitly specifies that the elements can be of any type. While the first can be better defined after further code analysis, the second can not.