Towards full integration with ADODB Data Dictionary. Step 3: ADOConnection::MetaType(), ADORecordset::MetaType(),  ADOConnection::$concat_operator, ADOConnection::sysDate, ADOConnection::sysTimeStamp, ADOConnection::$sysUTimeStamp, ADOConnection::SQLDate().
Given apparent specification, ADORecordSet::$connection must always be set to a valid ADOConnection instance. The few drivers that violated the specification were fixed.

The function ADORecordSet::SerializableRS() was in violation of the previously mentioned apparent specification, and fixed.

The function ADOConnection::newADORecordSet() was introduced, and code must call this function to create instances of ADORecordSet.

ADORecordSet::$datetime is not used anywhere in ADOdb and was removed. In the past the variable served a purpose in ADORecordSet::MetaType(). About 16 years ago, the variable was moved implicitly to ADOConnection, but was never declared there. Only one driver declared it, "adodb-oci8.inc.php".

The odbpt driver can no longer be used for DB2 and MySql in this fork. The implementation was incomplete, and incomplete enough to make it broken. However I intend to add support for DB2 and MySql if requested.

Some incorrect access levels of class variables were fixed. These are errors introduced in this fork after the adoption of access specifiers.

The variable ADODB_pdo::$stmt was not used anywhere in the ADOdb code, and appears to have been meant to be the implicitly declared variable ADODB_pdo::$_stmt. ADODB_pdo::$stmt was removed.

A minor bug in adodb-postgres64.inc.php implementation of ADOConnection::_connect() pertaining to ADODB_postgres64::$noBlobs was fixed.

Remnants of ADOConnection::SQLDate(), ADOConnection::$sysDate, and ADOConnection::$sysTimeStamp were removed from the Sqlite drivers, finishing that part of the work pertaining to the full integration with ADOdb Data Dictionary.

The driver "adodb-pdo_sqlite.inc.php" was updated to integrate the PHP functions ::adodb_date() and ::adodb_date2() into the sql, to keep to the behavior of the other sqlite drivers.