Towards full integration with ADODB Data Dictionary. Step 3: ADOConnection::MetaType(), ADORecordset::MetaType()
It appears that ADORecordset::MetaType() came first to translate driver specific types to ADOdb types. Later, the data dictionaries came, and it was decided convenient that the MetaType() function be available on them giving birth to ADODB_DataDict::MetaType(). This was a mistake because these functions are about the driver, yet their output is tied to the specification of dbTable::addField() which is part of ADOdb Schema which itself is the apparent reason for why the ADODB Data Dictionary was made, hence the whole confusion. Given the mistake, it was no surprise that later the implementation of ADODB_DataDict::MetaType() had to check first if a driver existed on the ADODB_DataDict instance before calling ADOConnection::MetaType(), and if not, supplying its own result and exasperating the problem instead of solving the original mistake. About a year later "adodb-replicate.inc.php" was introduced to ADOdb and made direct use of ADODB_DataDict::MetaType(). As for ADOConnection::MetaType(), it itself would create an internal instance of ADORecordset just to call ADORecordset::MetaType(), a product of another problem related to the code architecture.

ADORecordset::MetaType() implementations have been moved to the corresponding ADOConnection implementations. ADORecordset::MetaType() is still a valid function, and simply forwards the call to ADOConnection::MetaType(). As a consequence, ADORecordset::blobSize was also removed, and set on ADOConnection instead. This breaks backward compatibility, and user code should switch to using ADOConnection::blobSize instead.

ADODB_DataDict::MetaType() implementations have been removed, and so was ADODB_DataDict::blobSize. This affects no code in ADOdb except "adodb-replicate.inc.php". ADODB_DataDict::MetaType() is still a valid function, and simply forwards the call to ADOConnection::MetaType().

The former ADODB_DataDict::blobSize, is set by the ADOdb Data Dictionary for firebird. The value is kept for reference, but it would have only affected "adodb-replicate.inc.php". Hence, the value was not moved to the "adodb-firebird.inc.php" driver.

A fix was made to the resolution to the ADOdb Data Dictionary for the frontbase database, leading to the introduction of the "datadict-frontbase.php" data dictionary. This pertained to the "adodb-fbsql.inc.php" driver.

An error was fixed in the SQLite Data Dictionary pertaining to the previous commit.

An error was fixed in "adodb-replicate.inc.php" pertaining to the migration to PHP v5 work that was done in this fork.