Postliminary Commit(2)
The fix introduced to the "adodb-mysqli.inc.php" driver pertaining to "bound variables statements" is ill defined. However, the code continued to be useful for a proper fix. Thanks to the community. The stability of the solution was not confirmed "adodb-mysqli.inc.php". To revert to the old behavior prior to this commit and the pertaining commit, set ADOConnection::_bindInputArray to false in the source code of the driver, "adodb-mysqli.inc.php". With ADOConnection::_bindInputArray remaining true, the driver should revert to using the native driver functionality for preparing statements, and ADOConnection::Prepare() should be working correctly.

The old memcache implementation was brought back under the file "adodb-memcachelegacy.lib.inc.php". To revert to using the old implementation delete "adodb-memcache.lib.inc.php" and rename "adodb-memcachelegacy.lib.inc.php" to "adodb-memcache.lib.inc.php". Currently, there are no plans to integrate the old memcache system directly.

Decided to stick with the change of the interpretation of the parameter, $primary, of the function ADOConnection::MetaIndexes() for the driver "adodb-sqlite3.inc.php" that was made by the pertaining commit. The change in interpretation has also been applied to "adodb-sqlite.inc.php", "adodb-pdo_sqlite.inc.php", the odbtp_mssql driver in "adodb-odbtp.inc.php", "adodb-oci8.inc.php", "adodb-pdo_oci.inc.php", and "adodb-odbc_db2.inc.php". With this, all drivers should be compliant to this interpretation. Refer to the notes of ADOConnection::MetaIndexes().

The fix made to ADOConnection::MetaIndexes() for the driver "adodb-sqlite3.inc.php" by the pertaining commit has been applied to "adodb-sqlite.inc.php" and "adodb-pdo_sqlite.inc.php".

"perf-sqlite3.inc.php"  was renamed to "perf-sqlite.inc.php"