Pertaining to ADOConnection::qstr()
The function ADOConnection::qstr() can not be defined in terms of ADOConnection::addq() in a useful manner. An example is a string whose value is NULL when using MySQL (see "adodb-mysql.inc.php"). The example:

    $string = NULL;
    $qString = $db->addQ($string);
    $SQL = "SELECT * FROM names WHERE name='$qString'";
    $result = $db->execute($SQL);

paraphrased from the official documentation (https://adodb.org/dokuwiki/doku.php?id=v5:reference:connection:addq) would not work. No fix to the MySQL drivers can make it work.

ADOConnection::addq() is now final.

A fix was made to ADOConnection::qstr() and ADOConnection::addq() in the base class pertaining to the way they handled the ini setting, magic_quotes_sybase. The fix was deduced from other drivers, the git history, and references around the web, but was not tested.